# perf(game): add composite indexes for game_id and achievements unlocked