--- a/database/migrations/platform/2023_10_29_000000_update_player_games_table.php
+++ b/database/migrations/platform/2023_10_29_000000_update_player_games_table.php
@@ -10,16 +10,16 @@ return new class() extends Migration {
     public function up(): void
     {
         Schema::table('player_games', function (Blueprint $table) {
-            $table->index(['game_id', 'achievements_unlocked'], 'game_id_achievements_unlocked_index');
-            $table->index(['game_id', 'achievements_unlocked_hardcore'], 'game_id_achievements_unlocked_hardcore_index');
+            $table->index(['game_id', 'achievements_unlocked']);
+            $table->index(['game_id', 'achievements_unlocked_hardcore']);
         });
     }
 
     public function down(): void
     {
         Schema::table('player_games', function (Blueprint $table) {
-            $table->dropIndex('game_id_achievements_unlocked_index');
-            $table->dropIndex('game_id_achievements_unlocked_hardcore_index');
+            $table->dropIndex('player_games_game_id_achievements_unlocked_index');
+            $table->dropIndex('player_games_game_id_achievements_unlocked_hardcore_index');
         });
     }
 };