This dramatically improves performance on the two queries to fetch the game's achievement distribution on game pages.

**Benchmark query:**
```sql
-- softcore mode, use NOW() to bypass local performance caching
SELECT
	NOW(),
	pg.achievements_unlocked AS AwardedCount,
	COUNT(*) AS NumUniquePlayers
FROM
	player_games AS pg
WHERE
	pg.game_id = 228
	AND pg.achievements_unlocked > 0
GROUP BY
	AwardedCount
ORDER BY
	AwardedCount DESC
```

**Before**
4.389s
4.697s
4.282s
4.723s
4.867s
Average: _4.591s_

**After**
38ms
23ms
8ms
24ms
19ms
Average: _22.4ms_

The addition of these indices results in a speedup factor of around 535x. This is especially desirable given the query **runs twice on every game page**.