@Jamiras 2023-10-30
I'm not sure if the NOW() is enough to bypass my local caching.

> Before: 69ms, 71ms, 67ms, 70ms, 68ms (average: 69ms)
> After: 8ms, 9ms, 8ms, 8ms, 8ms (average: 8ms)

Still a 8x improvement.

For 1446:

> Before: 149ms, 78ms, 116ms, 94ms, 230ms (average: 133ms)
> After: 5ms, 5ms, 8ms, 7ms, 6ms (average: 6ms)

That's a 22x improvement