@luchaos 2023-10-30
> runs twice on every game page

Should probably be pre-aggregated on games/sets. Won't scale well in the future.